requirements.txt
Pillow==3.4.1
PyYAML==3.12
arrow==0.8.0
beautifulsoup4==4.5.1
blessings==1.6
chardet==2.2.1
colorama==0.2.5
cssselect==0.9.2
curtsies==0.2.9
decorator==4.0.10
feedfinder2==0.0.4
feedparser==5.2.1
futures==3.0.5
html5lib==0.999
idna==2.1
jieba3k==0.35.1
lxml==3.6.4
newspaper3k==0.1.7
nltk==3.2.1
pymongo==3.3.0
python-dateutil==2.5.3
requests==2.11.1
requests-file==1.4
six==1.10.0
tldextract==2.0.1
tomorrow==0.2.4
urllib3==1.7.1
validators==0.11.0
wcwidth==0.1.7
wheel==0.24.0
